pyproject.toml
[project]
name = "journal-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.51.0",
    "baml-py>=0.210.0",
    "elasticsearch==8.12.0",
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "openai>=1.70.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "uvicorn>=0.34.2",
]
